editor for testimonials with sanitize, and uploading avatars

+++ contollers/adminController.js
@@ -2,6 +2,7 @@ const jwt = require('jsonwebtoken');
 const bcrypt = require('bcrypt');
 const path = require('path');
 const sharp = require('sharp');
+const sanitizeHtml = require('sanitize-html');
 // Set up the client
 require('dotenv').config();
 
@@ -89,32 +90,64 @@ const updateWorkoutPlansPrice = async (req, res) => {
 // @method post
 const createTestimonial = async (req, res) => {
   const { fullname, profession, textBs, textEn } = req.body;
-  const { buffer, originalname } = req.file;
-  const filename = Date.now() + path.extname(originalname);
-  // store avatar img
-  const avatarPath = `images-${filename}`;
+  let avatarPath = '';
+
+  // sanitize HTML input
+  const config = {
+    allowedTags: sanitizeHtml.defaults.allowedTags.concat([
+      'a',
+      'span',
+      'div',
+      'p',
+    ]), // Allow default tags plus "a", "span", "div", "p"
+    allowedAttributes: {
+      '*': ['style'], // Allow "style" for all tags
+      a: ['href'], // Allow "href" for "a" tags
+    },
+    allowedStyles: {
+      '*': {
+        // For all tags
+        color: [/.*/], // Allow the "color" style, accept any value
+        'font-size': [/.*/], // Allow "font-size" style, accept any value
+      },
+    },
+  };
+  const bs = sanitizeHtml(textBs, config);
+  const en = sanitizeHtml(textEn, config);
+
+  if (req.file) {
+    const { buffer, originalname } = req.file;
+    const filename = Date.now() + path.extname(originalname);
+    // store avatar img
+    avatarPath = `images-${filename}`;
+
+    try {
+      // optimize image for avatar
+      const avatarImage = await sharp(buffer)
+        .resize(100, 100)
+        .jpeg({ mozjpeg: true, quality: 30 })
+        .toBuffer();
+
+      // upload avatar
+      fs.writeFile(avatarPath, avatarImage, (err) => {
+        if (err) throw err;
+        console.log('Image uploaded successfully!');
+      });
+    } catch (err) {
+      console.error(err);
+      return res.status(500).json({ message: err.message });
+    }
+  }
 
   try {
-    // optimize image for avatar
-    const avatarImage = await sharp(buffer)
-      .resize(100, 100)
-      .jpeg({ mozjpeg: true, quality: 30 }) // https://cyclic-muddy-hoodie-ox-sa-east-1.s3.sa-east-1.amazonaws.com/images/1690447140955.jpg
-      .toBuffer();
-
-    // upload avatar
-    fs.writeFile(avatarPath, avatarImage, (err) => {
-      if (err) throw err;
-      console.log('Image uploaded successfully!');
-    });
-
     // Create testimonial
     const testimonial = await Testimonial.create({
       fullname,
       profession,
       avatar: avatarPath,
       text: {
-        bs: textBs,
-        en: textEn,
+        bs,
+        en,
       },
     });
 
@@ -158,16 +191,41 @@ const getTestimonial = async (req, res) => {
 // @route  /api/admin/testimonials/:id
 // @method put
 const updateTestimonial = async (req, res) => {
-  try {
-    const { id } = req.params;
-    const { fullname, profession, text } = req.body;
+  const { id } = req.params;
+  const { fullname, profession, text } = req.body;
+  let avatarPath = '';
 
+  if (req.file) {
+    const { buffer, originalname } = req.file;
+    const filename = Date.now() + path.extname(originalname);
+    // store avatar img
+    avatarPath = `images-${filename}`;
+
+    try {
+      // optimize image for avatar
+      const avatarImage = await sharp(buffer)
+        .resize(100, 100)
+        .jpeg({ mozjpeg: true, quality: 30 })
+        .toBuffer();
+
+      // upload avatar
+      fs.writeFile(avatarPath, avatarImage, (err) => {
+        if (err) throw err;
+        console.log('Image uploaded successfully!');
+      });
+    } catch (err) {
+      console.error(err);
+      return res.status(500).json({ message: err.message });
+    }
+  }
+  try {
     // Create testimonial
     const updatedTestimonial = await Testimonial.findByIdAndUpdate(
       id,
       {
         fullname,
         profession,
+        avatar: avatarPath,
         text,
       },
       { new: true }
@@ -197,7 +255,7 @@ const deleteTestimonial = async (req, res) => {
   const { avatar } = req.body;
 
   try {
-    await fs.unlinkSync(avatar);
+    if (avatar) await fs.unlinkSync(avatar);
     await Testimonial.deleteOne({ _id: id });
     res.status(200).json({ message: 'Testimonial deleted successfully!' });
   } catch (error) {
